+++ src/components/App.js
@@ -3,7 +3,7 @@ import Header from './Header';
 import Main from './Main';
 import useSocket from '../hooks/useSocket';
 import GlobalStyles from '../styles/globalStyles';
-import BackgammonBoard from './BackgammonBoard';
+import BackgammonBoard from './BoardUI/BackgammonBoard';
 
 function App() {
     const [socket, isConnected, isConnecting] = useSocket("ws://localhost:3001");
